Move useFocusOnTransition and useToggleRole to internal/Panels folder
Enact-DCO-1.0-Signed-off-by: Juwon Jeong (juwon.jeong@lge.com)
